+++ integration-tests/gatsby-source-wordpress/__tests__/index.js
@@ -60,6 +60,10 @@ describe(`[gatsby-source-wordpress] Run tests on develop build`, () => {
   let gatsbyDevelopProcess
 
   beforeAll(async () => {
+    if (process.env.SKIP_BEFORE_ALL) {
+      return
+    }
+
     if (!isWarmCache) {
       await gatsbyCleanBeforeAll()
     }
@@ -105,7 +109,10 @@ describe(`[gatsby-source-wordpress] Run tests on develop build`, () => {
   require(`../test-fns/index`)
 
   afterAll(done => {
-    gatsbyDevelopProcess.kill()
+    if (gatsbyDevelopProcess) {
+      gatsbyDevelopProcess.kill()
+    }
+
     done()
   })
 })