+++ packages/gatsby/index.d.ts
@@ -17,7 +17,7 @@ import { GraphQLOutputType } from "graphql"
 import { PluginOptionsSchemaJoi, ObjectSchema } from "gatsby-plugin-utils"
 import { IncomingMessage, ServerResponse } from "http"
 
-export type AvailableFeatures = never // "image-service"
+export type AvailableFeatures = "image-cdn"
 
 export {
   default as Link,